requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.11.2
pydantic-settings==2.8.1
python-dateutil==2.9.0.post0
pytest==8.3.5
httpx==0.28.1
google-api-python-client==2.169.0
google-auth==2.40.1
